added extended euclidean
