--- Algorithms/CPP/bubblesort.cpp
@@ -1,34 +1,0 @@
-#include<iostream>
-using namespace std;
-  int n;
-void bubblesort(int arr[],int n);
-int main()
-{
-
-    cin>>n;
-    int arr[n];
-    for(int i=0;i<n;i++)
-    {
-        cin>>arr[i];
-    }
-    bubblesort(arr,n);
-    for(int i=0;i<n;i++)
-    {
-        cout<<arr[i];
-    }
-}
-void bubblesort(int arr[],int n)
-{
-    for(int i=0;i<n-1;i++)
-    {
-        for(int j=0;j<n-i;j++)
-        {
-            if(arr[j]>arr[j+1])
-            {
-                int temp=arr[j];
-                arr[j]=arr[j+1];
-                arr[j+1]=temp;
-            }
-        }
-    }
-}
