+++ Algorithms/JAVA/DetectLoopInLinkedList.java
@@ -1,0 +1,60 @@
+
+class LinkedList {
+    Node head; 
+    class Node {
+        int data;
+        Node next;
+        Node(int d)
+        {
+            data = d;
+            next = null;
+        }
+    }
+
+    public void push(int new_data)
+    {
+        /* 1 & 2: Allocate the Node &
+                Put in the data*/
+        Node new_node = new Node(new_data);
+ 
+        /* 3. Make next of new Node as head */
+        new_node.next = head;
+ 
+        /* 4. Move the head to point to new Node */
+        head = new_node;
+    }
+ 
+    void detectLoop()
+    {
+        Node slow_p = head, fast_p = head;
+        int flag = 0;
+        while (slow_p != null && fast_p != null
+               && fast_p.next != null) {
+            slow_p = slow_p.next;
+            fast_p = fast_p.next.next;
+            if (slow_p == fast_p) {
+                flag = 1;
+                break;
+            }
+        }
+        if (flag == 1)
+            System.out.println("Loop found");
+        else
+            System.out.println("Loop not found");
+    }
+ 
+    /* Driver program to test above functions */
+    public static void main(String args[])
+    {
+        LinkedList llist = new LinkedList();
+ 
+        llist.push(20);
+        llist.push(4);
+        llist.push(15);
+        llist.push(10);
+ 
+        llist.head.next.next.next.next = llist.head;
+ 
+        llist.detectLoop();
+    }
+}
